ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
sentence-transformers==2.3.1
rapidfuzz==3.6.1
httpx==0.26.0
pydantic==2.5.3
asyncpg==0.29.0
python-dotenv==1.0.0
pytest==7.4.4
hypothesis==6.92.2
scikit-learn==1.4.0
numpy==1.26.3
